review_agents_claude/java-review-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "java-review-agent"
version = "0.1.0"
description = "Java Code Review AI Agent using LangGraph and Ollama"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-community>=0.3.0",
    "ollama>=0.3.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
java-review = "java_review_agent.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/java_review_agent"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
